+++ src/app/gui/params/params.component.ts
@@ -3,6 +3,7 @@ import {MatDialog} from '@angular/material/dialog'
 import {ShowCodeDialog} from "../show-code-dialog/show-code-dialog.component"
 import {Params} from "@angular/router"
 import {DEFAULT_PARAMS} from "../../model/component-params"
+import {AddAnimationDialog} from "../add-animation-dialog/add-animation-dialog.component";
 
 @Component({
   selector: 'app-params',
@@ -16,10 +17,13 @@ export class ParamsComponent implements OnInit {
   public customColor = "red"
   public customBackgroundColor = "red"
   public customBackgroundImage: string
+  public animationList: string[] = ['slide-right', 'slide-in-from-left', 'fade-in-out']
+
 
   constructor(
     public dialog: MatDialog
-  ) {}
+  ) {
+  }
 
   get animationParamsKeys(): string[] {
     return Object.keys(this.animationParams)
@@ -33,7 +37,7 @@ export class ParamsComponent implements OnInit {
 
   private setAnimationParamsListener(): void {
     // Listen to answer event from animation
-    window.addEventListener('selected-animation-params', (event: any)=> {
+    window.addEventListener('selected-animation-params', (event: any) => {
       this.animationParams = event.detail
       for (const key in this.animationParams) {
         if (!this.params.hasOwnProperty(key)) {
@@ -98,5 +102,56 @@ export class ParamsComponent implements OnInit {
       exitAnimationDuration: 0,
       data: paramsToDisplay
     })
+  }
+
+  openAddAnimationDialog(): void {
+    const dialogRef = this.dialog.open(AddAnimationDialog, {
+      width: '80vw',
+      enterAnimationDuration: 0,
+      exitAnimationDuration: 0,
+      data: {}
+    })
+    dialogRef.afterClosed()
+      .subscribe((data: any): void => {
+        if (data) this.onCustomAnimationAdded(data)
+          .then((animationCode: string) => this.getAnimationName(animationCode))
+          .then((animationName: string) => this.addCustomAnimationToDropdown(animationName))
+          .then((animationName: string) => this.selectCustomAnimation(animationName))
+      })
+    // TODO: show error in toast if animation is not valid?
+  }
+
+  // TODO: maybe add animations from file with a file upload?
+  onCustomAnimationAdded(data: string): Promise<string> {
+    const animationCode = data.replace("export {}", "")
+    if (!animationCode.includes("registerAnimator") || !animationCode.includes("animationName")) {
+      return Promise.reject(null)
+    } else {
+      eval(animationCode)
+      return Promise.resolve(animationCode)
+    }
+  }
+
+  getAnimationName(animationCode: string): Promise<string> {
+    try {
+      let animationName = animationCode.split("animationName = '")[1]
+      animationName = animationName.split("';")[0]
+      return Promise.resolve(animationName)
+    } catch (e) {
+      return Promise.reject(null)
+    }
+  }
+
+  addCustomAnimationToDropdown(animationName: string): Promise<string> {
+    this.animationList.unshift(animationName)
+    return Promise.resolve(animationName)
+  }
+
+  selectCustomAnimation(animationName: string): void {
+    if (!animationName) return
+    this.params.animation = animationName
+    this.onUpdate()
+    this.getAnimationParams()
+    this.restartAnimation()
   }
 }