+++ src/app/shared/utils.ts
@@ -10,3 +10,5 @@ export const copyToClipboard = (text) => {
 export const generateId = (): string => {
   return 'animated-banner-' + Math.random().toString(36).substr(2, 9)
 }
+
+export const wait = (ms: number) => new Promise(res => setTimeout(res, ms))